// client/src/constants/Endpoints.tsx
export enum Endpoints {
    ROOT = "http://localhost:8000/interview/",
    CONTENT = "http://localhost:8000/interview/v1/content/topics/undefined",
    TOPICS = "http://localhost:8000/interview/v1/content/topics",
    HOTKEYS = "http://localhost:8000/interview/v1/hotkeys"
}

export enum QueryKeys {
    ROOT = "ROOT",
    CONTENT = "CONTENT",
    TOPICS = "TOPICS",
    HOTKEYS = "HOTKEYS"
}